# openapi.yaml
openapi: 3.0.3
info:
  title: PalletMS API
  description: Pallet tracking API with RBAC, movements, and reporting
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    CreatePalletRequest:
      type: object
      required: [barcode, currentAreaId]
      properties:
        barcode:
          type: string
        type:
          type: string
        size:
          type: string
        currentAreaId:
          type: integer
        owner:
          type: string
    CreateMovementRequest:
      type: object
      required: [palletId, toAreaId]
      properties:
        palletId:
          type: integer
        toAreaId:
          type: integer
        eta:
          type: string
          format: date-time
        notes:
          type: string
    ConfirmMovementRequest:
      type: object
      properties:
        notes:
          type: string
paths:
...
  /pallets:
    get:
      summary: List pallets (paginated, filterable)
      parameters:
        - name: barcode
          in: query
          schema: { type: string }
        - name: currentAreaId
          in: query
          schema: { type: integer }
        - name: conditionStatus
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
    post:
      summary: Create pallet
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePalletRequest'
...
  /movements:
    get:
      summary: List movements (paginated, filterable)
    post:
      summary: Start movement (Move Out)
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMovementRequest'
  /movements/{id}/confirm:
    post:
      summary: Confirm movement (Move In)
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmMovementRequest'
...